functions/requirements.txt
requests>=2.28
feedparser
beautifulsoup4
python-dotenv
firebase-admin
google-genai
python-telegram-bot
Pillow
langchain
langgraph
functions-framework
